switched to forked deps
